+++ ui/src/components/CuiAutomation/config/configMap.ts
@@ -12,6 +12,7 @@ export const CONFIG_MAP: Record<string, Component> = {
   'action-snapshot': asyncComponent(() => import('./ConfigActionSnapshot.vue')),
   'action-sensor': asyncComponent(() => import('./ConfigActionSensor.vue')),
   'action-notification': asyncComponent(() => import('./ConfigActionNotification.vue')),
+  'action-notification-control': asyncComponent(() => import('./ConfigActionNotificationControl.vue')),
   'action-http': asyncComponent(() => import('./ConfigActionHttp.vue')),
   'action-mqtt': asyncComponent(() => import('./ConfigActionMqtt.vue')),
   'action-delay': asyncComponent(() => import('./ConfigActionDelay.vue')),